+++ .gitmodules
@@ -1,3 +1,4 @@
 [submodule "gz"]
 	path = gz
 	url = git@github.com:valandil/gz.git
+
