+++ n64split/configs/sm64.shindou.yaml
@@ -2943,7 +2943,7 @@ labels:
    - [0x8028F48C, "proc_8028F48C"]
    - [0x8028F4D8, "proc_8028F4D8"]
    - [0x8028F61C, "proc_8028F61C"]
-   - [0x8028FA68, "proc_8028FA68"]
+   - [0x8028FA68, "DistanceFromObject"]
    - [0x8028FAAC, "proc_8028FAAC"]
    - [0x8028FC04, "proc_8028FC04"]
    - [0x8028FC98, "proc_8028FC98"]
@@ -4163,8 +4163,10 @@ labels:
    - [0x80309900, "__osSiDeviceBusy"]
    - [0x8030A350, "__osSetCompare"]
    - [0x8030A360, "__osProbeTLB"]
+   - [0x8031D9C0, "gMarioStates"]
    - [0x8033CD14, "gPlayer1Controller"]
    - [0x8033CD1C, "gPlayer2Controller"]
+   - [0x80343320, "gCurrentObject"]
    - [0x80378800, "Copy32BitTriple"]
    - [0x8037882C, "proc_8037882C"]
    - [0x8037885C, "proc_8037885C"]
